Support nullability improvements for SELECT table.*
Summary: We already supported nullability improvements for the `SELECT *` case. This adds support for `SELECT table.*`. It also fixes a crash that would occur if one attempted to improve a column from a `table.*` expression with an id (e.g., `select t.* from t where x is not null`) as opposed to a dot (e.g., `select t.* from t where t.x is not null`).

Reviewed By: ricomariani

Differential Revision: D31099200

fbshipit-source-id: 2caed5e1f42a3e74c3d5258244c1ae310e3c5eae