Fix query plan generation for object types
Summary:
In `variables_callback`, we were generating `'1'` as a stand-in parameter for object types with the assumption that, in the original code, the object parameter being replaced would always be passed to `ptr`, and thus we needed a numeric type to replace the whole expression as `ptr` was not being emitted into the query plan output. This meant that select functions that required a true object type (as opposed to a `LONG` from `ptr`) were being passed `'1'` and then would fail to type check.

This fixes the query plan output by emitting `cast('1' as object)` as the stand-in for object types and by changing the callbacks mode to `gen_mode_no_annotations` so that functions like `ptr` are emitted into the query plan output (as they are critical for analysis of the output to type check).

Reviewed By: ricomariani

Differential Revision: D31318564

fbshipit-source-id: f11fea7cd853738078deee31a07cf3977e3789fc