Add NULL CHECK ON NOT NULL enforcement option
Summary:
There is generally no reason to perform an `IS NULL` or `IS NOT NULL` check on a value of a `NOT NULL` type. Doing so often indicates confusion that may lead to unexpected behavior (e.g., checking, incorrectly, if a cursor has a row via `cursor IS NOT NULL`). For the rare cases where one really does want to check if a value of a `NOT NULL` type is or isn't null -- cases for which the need will be eliminated in the future, hopefully -- one can always use `nullable`.

Enforcement for this will soon be the default: The temporary presence of `NULL CHECK ON NOT NULL` is merely to enable the conversion of existing code.

Reviewed By: RaoulFoaleng

Differential Revision: D31669020

fbshipit-source-id: c256b021009918bc3b9012dbdceea3d49d234d15