add min version argument to upgrader
Summary: If you don't need to be able to upgrade from the distant past, this can give you a path for more recent peoples.  The idea is that you can ship say a current empty db shapshot for new instances and upgrade anyone who already has a db that is recent enough.

Reviewed By: winniequinn

Differential Revision: D31905681

fbshipit-source-id: c56831a799e7d2fecdf0173a37451a73b9f7d5db