Fix nullability and sensitivity for substr
Summary: The `substr` function returns null if any of its arguments are null. This fixes the analysis of `substr` such that it produces a nullable result if any of its arguments are nullable, not merely if the input string is nullable. It also changes the sensitivity analysis in the same way; the fact that a sensitive argument is recoverable by comparing the input and output strings alone implies that the output should be considered sensitive if any of the arguments are.

Reviewed By: ricomariani

Differential Revision: D32004221

fbshipit-source-id: 43af651870b44aa866b08c699e682b7d602641e1