Make PROC AS FUNC ARGUMENTS enforcement mandatory
Summary: It was always the plan that the presence of the `PROC AS FUNC ARGUMENTS` enforcement option would be a temporary situation to allow soon-to-be-dead code to be removed in a separate step. This change removes said dead code (including several now-irrelevant tests), updates documentation accordingly, and removes the enforcement option.

Reviewed By: ricomariani

Differential Revision: D32077880

fbshipit-source-id: 8626aeaf21a6cbd0477a6d904c6a392fc65978ee