Add PROC AS FUNC ARGUMENTS enforcement
Summary:
Procedures with `OUT` or `INOUT` parameters before the final `OUT` parameter should not be used as functions. Using them as functions not only results in confusing code, it also complicates analyses for things like nullability improvements (see `sem_unset_notnull_improvements_for_out_args`).

This enforcement disallows the use of a procedure as a function unless it has exactly zero or more `IN` parameters followed by exactly one `OUT` parameter. This new enforcement is initially optional to allow time for existing code to be converted, to allow "sem_test.sql" to continue to contain tests like `improvements_respect_out_args_in_true_conditions` until all code is converted, and to allow simplification of code in "sem.c" that currently must account for the proc-as-func-with-non-trailing-OUT-parameters case in a later step.

Reviewed By: ricomariani

Differential Revision: D32027878

fbshipit-source-id: b1960f6d0138d4082268037d7f09894c89ca664e