Require that OUT/INOUT args be unique per call
Summary:
Using the same variable as more than one argument in a call can lead to unexpected or unsafe behavior if one of those arguments is being passed for an `OUT` or `INOUT` parameter. It's easy to see how confusion could result if the same variable were passed for multiple `OUT` or `INOUT` parameters, but even passing a variable for a single `OUT` parameter and an `IN` parameter is unsafe. For example:

```
create proc p(a text not null, out b text not null)
begin
  set b := "goodbye";
end;

create proc main()
begin
  let t := "hello";
  call p(t, t);
end;
```

If we compile and execute this program, we'll actually crash as soon as `p` is called. The reason is that `main` will set `t` to `NULL` before the call (`cql_set_string_ref(&t, NULL);`) because the second parameter of `p` is an `OUT` parameter, and then we'll crash upon entering `p` because the first argument is not nonnull as expected (`cql_contract_argument_notnull((void *)a, 1);`).

While it is possible that we could generate code differently to avoid this particular problem, it's arguably a questionable idea to use the same variable multiple times within the same call if it is used as an `OUT` or `INOUT` argument regardless, and we'd still want to disallow using the same variable as multiple `OUT` or `INOUT` arguments to avoid surprising mutations. We, therefore, now simply disallow non-unique uses of variables passed as `OUT` or `INOUT` arguments altogether. (In the future, if desirable, we can relax this enforcement to ignore `IN` uses after the code generation has been updated such that doing so would no longer be unsafe.)

Note that it is perfectly fine to use the same variable within a subexpression of another argument, and we do indeed allow that case. As per the new documentation:

```
CREATE PROC some_proc(IN a TEXT, OUT b TEXT)
BEGIN
  ...
END

DECLARE t TEXT;

-- This is NOT legal.
CALL some_proc(t, t);

-- This, however, is perfectly fine.
CALL some_proc(some_other_proc(t), t);
```

The reason why this is always fine is that, when we emit C, we generate a temp for the result of `some_other_proc`, call it, and only then null out `t` and call `some_proc` with the previously mentioned temp as the first argument, not `t` as before.

Reviewed By: ricomariani

Differential Revision: D32111097

fbshipit-source-id: 87d519eed1b888cba120c4ef5a8a2a1cc518e3c1