Fix crash in sem_update_entry on lossy conversion
Summary:
It was the case that `sem_update_entry` would crash if there was a lossy conversion:

1. `sem_update_entry` called `sem_verify_assignment` with an `update_entry` as the first argument.
2. `sem_verify_assignment` called `sem_verify_safeassign`, again with the `update_entry` as the first argument.
3. In the error case, `sem_verify_safeassign` called `gen_with_callbacks` to print the `update_entry`.
4. There is no registered expression type for an `update_entry` in "gen_sql.c", so CQL crashed in `gen_expr` on the invariant.

This fixes the above by changing step 1 such that `sem_verify_assignment` is called with the problematic expression as the first argument (as is done in `sem_validate_compatible_table_cols_vals`) instead of the `update_entry`. A new test that previously would have crashed demonstrates that the bug has been fixed.

Reviewed By: ricomariani

Differential Revision: D32227198

fbshipit-source-id: 640a0ac1f11f3f1979c9fe4057531c826c34582a