Add exclusive mode schema support
Summary:
# What is Exclusive Mode

The `--rt schema_upgrade` code fundamentally assumes that it is not the only player in any particular database.  In fact we expected that you might want two different upgraders in the same db upgrading different/unrelated pieces of schema.  This is very flexible but it does come at some cost.  In particular you have to have a tombstone for every piece of schema, including views, indices, and triggers -- which are in some sense either ephemeral or recreatable.  They're never completely precious (though indices are expensive to recreate).

The idea here is to add an exclusive mode to the generator.  This tells it that the generator owns ALL of the schema for this database.

# What does this enable?

* when dropping all views we can simply enumerate from sqlite master and drop everything rather than have a list
* when dropping all triggers we can do the same
* when dropping indices, we want to be more careful but we can drop any index that isn't one we know to clean things up

All of these together mean that tombstone delete entries are not needed for correctness when making an upgrader in this mode which cleans up the schema somewhat.  However, if you do this, there is no going back... unless you can somehow restore the tombstones of everything you ever deleted.

The previous schema verifier is relaxed in this mode so that tombstones are not needed for views, tables, indices.

# How do I opt in?

You get in this mode by supplying `--schema_exclusive` when creating your schema or validating previous schema.

# Why not tables too?

The problem with no tombstones for tables is "reincarnation".  Suppose we have these steps:

* in v5 you create table foo
* in v6 you drop table foo
* in v7 you create foo again, but it's a different foo

This nearly works.  Your users get foo in v5, then in v6 its gone from the schema, so as an unknown table its deleted, just like if there was a tombstone then in v7 the new foo table is created again and all is well.

The problem is, what if some of your users never upgraded to v6?  Or supposed they had a backup of a database at v5?  Now they're in a bad place.  The schema had no tombstone for v5.  When it looks for the table foo it will find it but it's a different foo and it has no way to know that it is a foo from the past that should be deleted. The upgrader does not drop tables on the `create` plan so it's stuck.

Now if the new foo is on the `recreate` plan that could work... but this is now getting very hard to understand.

To fix this you sort of have to keep side-state of any table that used to exist but this is really problematic.  We're trying to make things easier by adding exclusive mode.  Putting the schema history knowledge in two places one of which is invisible to CQL is not likely to make things simpler in the long run -- it's just asking for problems.

So until we have some better idea, we just do this simplification for VIEWS, TRIGGERS, and INDICES.

# What else is in this diff?

The special kludge to drop all triggers that start with "tr__" has been removed.  It was a kludge for some very old CQL uses that no longer exist and should have been deleted years ago.  This seemed like a good time to remove that too.

Reviewed By: winniequinn

Differential Revision: D32221990

fbshipit-source-id: c0ff88ff8a9530725d18d14532dcb43547a8814c