CQL support declaring global constants
Summary:
This adds new statements:

```
declare const group foo (
  const_v = false,
  const_w = 3.5,
  const_x = 1L,
  const_y = 2+3,
  const_z = "hello, world\n"
);
```
and

```
emit_constants foo;
```

The idea is that you can define global constants which are usable anywhere.  They are analogous to declared `enums` but:

* they do not have a scope so you don't say `foo.const_v`, you just say `const_v`
* they do not have to have the same type
* just like enums they go into the .h file that contains `emit_constants` so you can use them from C

The code in the .h file file ends up looking like this:

```
#ifndef const_group_foo_defined
#define const_group_foo_defined

#define u 0
#define w 3.500000e+00
#define x _64(1)
#define y 5
#define z "hello, world\n"

#endif
```

These constants will soon go into the JSON output alongside enums so that they can be consumed by other parts of the system.

### Why do we need `emit_constants`?
Just like with enums, we expect that these will be pulled via `#include` into many contexts, so for instance if the constant block is in `foo.inc` you might see `#include "foo.inc"` in many contexts.  Say `foo.sql` and `bar.sql`.  Those two files normally would generate `foo.h` and `bar.h` as well as `foo.c` and `bar.c`.  We don't want the definitions for the constant values to appear in literally every header file.  So `foo.sql` would include `emit_constants foo;` causing its output header file `foo.h` to get those values.

### Why do we emit `#define` rather than const ?
If we emitted const we'd get a read-only `const int` variable or something like that for each constant.  We don't want that -- we want a compile time constant.  The only way to get that in C is to use `#define`.

### What's this going to look like in JSON?
TDB but you can look at what we do for enums and it will likely be almost exactly the same.  The main difference is that each constant needs a type.

### Why do we need a group name?
It's only for convenience when using `emit_constants` so you can define a bunch and emit them together.  The group name is otherwise not important.

Reviewed By: winniequinn

Differential Revision: D32292619

fbshipit-source-id: c9f341a9c0c9b661b42a0fb99ca3c980369ff151