Allow several variables named *database to contribute to the final JSON global attributes output
Summary:
Currently if you have various schema fragments you might find yourself in a situation where each fragment is suppose to contribute some useful attributes to the JSON that control your down-level systems.  But there can only be one global "database" object variable to put your global attributes on.

Here we extend the system so that any variable that ends with "database" and is of type object gets its attributes appended to the global attributes section in the JSON.

Docs updated to reflect the new reality.

Differential Revision: D32456697

fbshipit-source-id: c024b1e4feb50819c7ba6926f28e911c4bfa39d6