Make NULL CHECK ON NOT NULL a no-op
Summary: It was always the plan that `NULL CHECK ON NOT NULL` would be a temporary enforcement option and would soon become the new, non-negotiable default behavior. This changes CQL such that the enforcement is always in effect while continuing to parse (and ignore) the enforcement option for a brief window of backwards compatibility.

Reviewed By: ricomariani

Differential Revision: D32261191

fbshipit-source-id: f64b08dbd3651d47b01b3a3065ca637dc13eaf95