Error out on bad window definitions
Summary: We were printing errors in window definitions, but the lack of a check on the result of `sem_window_defn` in `sem_window_name_defn` was allowing semantic analysis to proceed and bad code to be generated. This fixes the problem by simply checking the result of `sem_window_defn` appropriately.

Reviewed By: ricomariani

Differential Revision: D32570841

fbshipit-source-id: 15cbaf123bcb80c2f11abf53ea4a76d166c60af6