Migrate nullability improvements
Summary:
This changes nullability improvements to use the new flow system. As a result, additional nullability improvements can now be made.

It is now the case that setting something to `NULL` in a branch no longer always results in it being nullable after the `IF`:

```
DECLARE a INT;

SET a := 42;

IF some_condition THEN
  SET a := NULL;
  SET a := 100;
ELSE
  ...
END IF;

-- `a` is still nonnull here despite the above set to NULL
```

Additionally, if all branches improve a variable, it is improved after the `IF` if the `IF` contains an `ELSE` branch:

```
DECLARE a INT;

IF some_condition THEN
  SET a := 42;
ELSE
  SET a := 42;
END IF;

-- `a` is nonnull here because all branches improved it
```

**NOTE:** For various reasons, the analysis for loops is not entirely sound, but it was not sound before this change either. This will be fixed in a follow-up.

Reviewed By: ricomariani

Differential Revision: D33012800

fbshipit-source-id: 7502f0e500864b31448ca3011fc7f6ee6f292647