Add jump contexts and flow analysis for loops
Summary:
Up until this point, improvements in CQL were unsafe within loops. Take the following code as an example:

```
001  DECLARE x TEXT;
002  SET x := "foo";
003  WHILE some_condition
004  BEGIN
005    CALL requires_text_notnull(x);
006    SET x := NULL;
007  END;
```

It is unsafe to allow the call to `requires_text_notnull` on line 5 despite `x` being improved on line 2 because `x` is later unimproved on line 6 and the statement list within the loop may execute more than once. See `unimprovements_in_loops_affect_earlier_statements` in "sem_test.sql" for an example involving nested loops.

Flow analysis improvements that merged the effects of branches recently introduced another problem:

```
001  DECLARE x TEXT;
002  SET x := "foo";
003  WHILE some_condition
004  BEGIN
005    IF another_condition THEN
006      SET x := NULL;
007      IF yet_another_condition THEN
008        LEAVE;
009      END IF;
010      SET x := "bar";
011    ELSE
012      -- do nothing
013    END IF;
014  END;
015  CALL requires_text_notnull(x);
```

Here, even though the outer `IF` is neutral with respect to the nullability improvement to `x` from line 2—it unsets it on line 6 and then re-sets it on line 10 and the `ELSE` does nothing—there is no guarantee that line 10 will ever be evaluated because we may jump straight from line 8 to line 15. See `loops_keep_all_unsets_and_ignore_all_sets` in "sem_test.sql" for another example.

This change solves both of the above problems. The first is solved by analyzing the loop *twice*. The latter is solved via a new form of flow context called a jump context which pessimistically assumes all unsets within it occurred and all sets within it did not occur.

**NOTE:** Exceptions thrown within a `TRY` block cause a problem similar to the second issue mentioned above. Jump contexts will be used to fix that as well in a later change.

Reviewed By: ricomariani

Differential Revision: D33143225

fbshipit-source-id: a479a9271928320c0de7b74245b151949bd7b4a4