Allow use of shared fragments in simple nested select statements
Summary:
You can now write

select * from (call shared_fragment(args));

The only tricky business about this is that we have emit a wrapper around the fragment when we do this because the fragment's column names are suppressed.  If we turned off column alias minification that would work too but then the text of the fragment would be different so it couldn't be shared.  So instead we make a local cte called `_ns_` for nested select which gives us the columns back.  It means a little extra text but not much and the extra text has the possibility of being shared.

Reviewed By: winniequinn

Differential Revision: D33178933

fbshipit-source-id: 00e86781aaa45165afc7afdc411fc21069b1ba15