Generate create_triggers()/drop_triggers() helpers even if empty
Summary:
We'll generate the create_triggers() and drop_triggers() helpers even for procedures
that do not use any tables with triggers.

This makes it very slightly more convenient to use these helpers in edge-cases where
the existence of triggers might change over time. It's possible to just add a call
to create_triggers() in a test and not worry about adding/removing triggers later.

Reviewed By: RaoulFoaleng, winniequinn

Differential Revision: D33146600

fbshipit-source-id: 462189e7cae05b5afc47a34cfda8bba001d41033