Shared fragments checks for parameters that are eclipsed by local CTEs
Summary:
This made it into the initial documentation but we didn't actually check the condition until now.

The issue is this if you do something like this

```
  (call shared_frag() using foo as source)
```

There is the possibility that `shared_frag` already has its own `foo`.  If it does then when
`source` is replaced with `foo` it will bind to the local `foo` which is not what you intended
and unlikely to even execute. So here we check to make sure that your parameter will not
conflict with an existing CTE in the named fragment.  Simple enough -- of course you could
have many such parameters, so they all have to be checked but more importantly `shared_fragment`
might forward your table parameter value to an inner shared fragment and that might have
a conflict.  As a result the shared fragments have to be recursively walked, but only
the chain that forwards an incoming parameter.

For simpler diagnosis the error message reports the chain of calls that resluted in the conflict if there is one
as well as the table parameter bindings that relevant.

Reviewed By: RaoulFoaleng

Differential Revision: D33247970

fbshipit-source-id: baea28d7962f42bcdbb36cb8cc0fd917a73328fd