Detect something that looks like it might be a doc comment before create proc, table, or view statements
Summary:
Any comment that starts with /** or /*! is converted to a misc attribute `cql:doc_comment` with the body of the comment as the value.  Any downstream tool can see this attribute in the JSON output and
use it to do whatever it likes.  This can be used for later stage codegen or for creating a doc compendium or whatever you might fancy.

Reviewed By: winniequinn

Differential Revision: D33282639

fbshipit-source-id: 767e89356cc08f36108b38b2143a7caa0bcb6964