Unify resolution for argument bundles
Summary:
For parameters within argument bundles, it was the case that `find_mutable_type` would return different type pointers depending on whether `bundle.parameter` or `bundle_parameter` was used. This meant that setting an improvement for  `bundle.parameter` would not result in it being set for `bundle_parameter`, and vice versa.

This unifies the resolution of parameters within argument bundles such that both forms now return the same type pointer. As a result, setting an improvement for one form of a reference to a parameter within an argument bundle now also also improves the other (as demonstrated in the new test `dot_form_and_underscore_form_are_equivalent`). This, in turn, lets us remove a rather ugly workaround in `sem_validate_current_proc_params_are_initialized` that was necessary for forwarding procedures.

Reviewed By: ricomariani

Differential Revision: D33313978

fbshipit-source-id: 2d80d5f13af2ceb0c7de816d81e7156b05d80dec