Towards --rt objc more general (#101)
Summary:
The code has assorted assumptions that are only good for the FB internal use case.  It should be more generic.

Note I may not have everything pulled out properly yet like I did with count/hash/copy so for now let's consider this a step in the right direction.  Further audits are needed.

Also coming soon: a `cqlrt_cf` that uses CFTypes for string/blob/etc instead of roll-your-own.  This is needed if you want objc support to really work because it assumes cql_string_ref can be converted to an NSObject with toll-free bridging.  That isn't the case for the generic version of the runtime.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/101

Test Plan: test options that work for both flavors will have to be added/revisted

Reviewed By: winniequinn

Differential Revision: D33340049

Pulled By: ricomariani

fbshipit-source-id: 7b84e1f1b11e9da26960c8629658b1340de04391