cqlrt_cf and a demo application to allow use of CF/objc with the OSS distro
Summary:
In response to this issue:

https://github.com/facebookincubator/CG-SQL/issues/99

It became obvious that there were some flaws in the objc code gen OSS version.  In particular the rt features that allow for name prefixing and various case conversions in the output were not used consistently which resulted in the OSS version of the system being broken, it wouldn't even compile.  That much has already been fixed.

Now comes the further problem that the normal cqlrt.c that we ship has no hope of interoperating with the Objective C Foundation.  This is not a problem with the internal versions of the runtime because they use CF friendly versions of the reference types.  In short in that version a cql_string_ref *is* a CFStringRef.

To make it possible to actually use the objc wrappers, this diff includes a new cqlrt flavor, cqlrt_cf which uses the CF for its reference management.  Cqlrt is designed to be replaced like this, you get to define what a string ref is, you just provide suitable helpers.

The cqlrt_cf directory includes a simple demo app that uses the objc friendly rowset accessors. It's very similar to the sample app that is now in appendix 10 of the guide (https://cgsql.dev/cql-guide/x10/)

Note that this app has no hope of running where CF isn't.  Hence it's not the portable choice (that's why the default cqlrt is the default).  But this does give us a simple example illustrating how cqlrt can be customized and it gives some insight into how Meta uses CQL on iOS which may be of interest.

With `--rt objc_mit` available we might be able to use that in `test.sh` rather than normalize the output instead of the `--test` aware code we use now  But there's no hurry there.

Reviewed By: winniequinn

Differential Revision: D33345934

fbshipit-source-id: 694b8191e525eefec5c7e388c87ddf450fb7ec2a