Fixed missing copyright headers... these were inadvertently removed; add cqlrt_cf reference in the docs
Summary:
Due to some rebase silliness the necessary headers were not present in the first instance in some files.

These are all covered by the same copyright as everything else, to be clear.

I have also added a brief reference to this new runtime as a sample in the internals documents on CQLRT.

Includes doc refresh script output.

Reviewed By: winniequinn

Differential Revision: D33370639

fbshipit-source-id: 4bc78e536c6a294d8be52acd5bf0212e06d62147