cqlrt_cf_ demo was built for ARC *off* which isn't normal... changed to ARC on
Summary:
It's not hard to switch it the other way but this is how everyone uses it anyway

Only CQLHolder really needs to change, and only bridge -> bridge_transfer.  The other
changes are to add a release pool block so cleanup is observable and fix the comments
which were actually wrong because I didn't understand ARC was off when I did this...

make.sh turns on ARC for this demo.

Reviewed By: wanaya

Differential Revision: D33392583

fbshipit-source-id: c1e256eccc97d0bb5d9d455f836198025c681fc7