migration procedures are auto-declared with the correct signature
Summary:
It's currently possible to declare that a proc is to be a migration proc with a version annotation but then when you actual create the proc the signature isn't correct.

To fix this we simply generate a synthetic declaration of the proc when you use it as a migration.  If the proc was previously declared something else, it will fail.  It if it is later declared other than the correct signature that also fails.

Reviewed By: RaoulFoaleng

Differential Revision: D33605909

fbshipit-source-id: 9c18dd23e8a69be718ccb5d0791c44c8ae95bfca