Support @delete of tables on the @recreate plan
Summary:
This deals with a long-existing error in the schema upgrader which was somewhat mitigated when we changed the drop strategy to drop all tables that are deleted rather than checksum the delete instructions.

The specific issue is this:

A table that was on the recreate plan was annotated like this:

```
create table dropping_this
(
  f1 integer,
  f2 text
) recreate(group);
```

When it is deleted, it was annotated like this:

```
create table dropping_this
(
  f1 integer,
  f2 text
) delete(5); -- whatever version
```

The problem is that the deleted version of `dropping this` is not distinguishable from a Baseline table (v0) that was subsequently deleted.  This transition was allowed and so `dropping_this` would be added to the Baseline schema (v0) and that schema, having changed its CRC would re-run.   This is not a huge problem since `dropping_this` would be deleted a few milliseconds later when the v5 deletes ran.

The problem is though, that `dropping_this` really is in the baseline schema now.  And if this happens again because say `dropping_that` is marked in the very same way then the baseline schema changes that time the v5 delete instructions for `dropping_this` won't run because the CRC will still match.  That leaves `dropping_this` in a zombie state.

Our previous change avoids this particular problem by always running the necessary drops, regardless of CRC.  However we still have a vestige in the baseline schema taking up space for no reason.

The presence change makes it possible/necessary to write:

```
create table dropping_this
(
  f1 integer,
  f2 text
) recreate(group) delete; -- no versions
```

This allows `dropping_this` to stay on the recreate plan and stay in its group.  As a result:

* no bogus baseline schema is created
* the table is dropped in the correct order in its group

Additional enforcement:

* deleted recreate table stays on the recreate plan
* recreate tables do not change groups

This ensures that the drop for the delete table is in the right place (in the recreate group) and there is no vestigial schema created in baseline.

Reviewed By: RaoulFoaleng

Differential Revision: D33804065

fbshipit-source-id: 69da238f1485562c44246fa7cdd91ca0a4d48b77