Support find_mutable_type for cursors
Summary:
It was the case that calling `find_mutable_type` on a cursor would always return NULL. This was because `sem_resolve_id` eventually resolved the cursor via `sem_try_resolve_cursor_as_expression` which set the type on the AST as though the cursor were used in an expression context (where it has a boolean type). Since `find_mutable_type` is used to look up a type *in the environment*, returning a pointer to the boolean type would have been useless, and so NULL was simply returned instead.

This moves things further in the direction of `sem_resolve_id` always resolving to a type in the environment with `sem_resolve_id_expr` handling the additional work inherent in resolving within an expression context. Now, only `sem_resolve_id_expr` sets a boolean type when resolving a cursor: `sem_resolve_id` simply sets the cursor type itself. This allows `find_mutable_type` to work correctly for cursors, thus enabling future work where improvements will be set on cursors themselves.

Reviewed By: ricomariani

Differential Revision: D33953737

fbshipit-source-id: dbd3567b988b5459be507709ad18713b59310f47