Add sign function enforcement
Summary:
The `sign` function is unavailable in versions of SQLite prior to 3.35. This adds an enforcement to disallow use of the function within SQL.

Note: While 3.35 added many new functions, `sign` is the only function added that can be used when SQLite is not compiled with `SQLITE_ENABLE_MATH_FUNCTIONS` defined. For this reason, no other enforcements are necessary and we can just target the `sign` function specifically. Should we run into another case like this, a more general means of enforcement might be appropriate.

Reviewed By: ricomariani

Differential Revision: D34437617

fbshipit-source-id: 00a9f75e21d79b95f5986a8db6e5cd617a0a9ae9