CQL structured blob storage -- v0.1
Summary:
There are many situations where you want to be able to create a blob of data that holds many different fields.  In order to do this economically and to use the type system the CQL provides you need just a two elementary operations.

* make a blob from a cursor
* make a cursor from a blob

Here we add the language notions to do this, they are akin to what we have for boxed cursors.

```
set my_blob from cursor my_cursor;
fetch my_cursor from my_blob;
```

To ensure type safety, the blob must carry a type discriminator (it's called "kind" in our docs) that matches a table with the appropriate shape and of course the cursor must also match.

Here's a simple example.

```
create table storage(
  x integer not null,
  y real not null,
  z long not null
);

create proc make_blob(out result blob<storage>)
begin
  declare C cursor for select 1 x, 3.5 y, 2L z;
  fetch C;
  set result from C;
end;

create proc make_cursor(stuff blob<storage>)
begin
  declare C cursor like storage;
  fetch C from stuff;
end;
```

For now any table can be used for the shape but this will need to be locked down.  Really, the shape must be devoid of table things like PK, FK, and so forth.  This will be enforced in subsequent checkins.

The blob that is created has the type encoded in it so extraction can fail if you provide a mismatched blob.  However, should the storage version so that it gains new nullable columns at the end, null will be provided for those missing columns.

The blob reading code is hardened so that it assumes the content of the blob is hostile.  There are tests for this.

The encoding is roughly:

* plain text metadata for the column types
* a bit vector for the isnull bits as needed and any booleans
* integers and longs are encoded in a variable length form using zigzag encoding just like gbuf
* floats are IEEE 754
* strings and blobs are stored inline
  * it's possible to nest structured storage

To avoid language cross product explosion and because field-at-a-time access to such a blob is very expensive the only operations are encode and decode.  Working with the blob requires that you break it into a value cursor.

Reviewed By: winniequinn

Differential Revision: D34429186

fbshipit-source-id: a3a39ad50d672cb318e13ce21d94d36dc61d64f2