Apply blob storage validations to tables
Summary:
In order to be used as blob storage a table has to be just storage.  It isn't a real table so the presence of FK, PK, constraints, etc.  Anything other than storage indicates a problem.

On columns, the only attributes that make sense are NOT NULL and `sensitive` and `create`. Anything else disqualifies the table for use as storage.

Also virtual tables cannot be storage.

Notes:

`delete` on columns seems reasonable at first, especially since you can only `delete` nullable columns anyway.  The problem is that if you were to delete a column then the decoder has to know about the column even though its been deleted and so that it can ignore the column if it is present in an older persisted blob.  It also means that the semantic type of the blob (which doesn't include the deleted columns) could not be used to create the serialization info.  This all seems to be a lot of work for a pretty confusing feature.  So we're left with the only versioning option as being to add new columns at the end.

`default value` is nominally doable but also results in a bit of potential confusion.  It could mean:
  * when loading a cursor of this shape, use the default value if one is missing [this seems doable]
  * when decoding a blob use the default value for the column if it is missing [this seems like a bad idea]

Both of these have consequences that are potentially error-prone in terms of expectations, so in the interest of simplicity and correctness neither is supported for now.  If it turns out to be necessary some choice could be made.  Probably the first alternative for default value.

Reviewed By: winniequinn

Differential Revision: D34620315

fbshipit-source-id: cbde1431fc5d8fe02a3b091c9d9a65e76552fbe1