Changing extension fragments to the "lite" strategy [breaking change for extension interfaces]
Summary:
This is a feature take-back of sorts, maybe the first.  It turns out the having a separate modular interface each extension in an assembly:

* complicates the codegen significantly
* isn't that useful
* prevents us from having several assemblies that use the same base but are otherwise independent

We can solve all these problems by simply not doing any codegen for the extension fragments, the assembly is the only "real" thing.  The base still defines a shared interface for the common columns.

There were very few uses of such extension interfaces and they've all been removed.  The way forward is to use the assemblies interface to access the columns you need, the assembly becomes the only "real" thing.

As a consequence we get to delete a lot of code and we have a path forward to support multiple extensions of a base which is actually a very useful feature.

*Future looking*

Independently of this, you can also expect that we'll simplify the extension syntax so that you can write them to look just like shared fragments.

Instead of:

```
attribute(cql:extension_fragment=base_frag)
create proc adds_columns(id_ integer not null)
begin
  with
    base_frag(*) as (select 1 id, "name" name, 1.0 rate),
    col_adder_frag(*) as (
    select base_frag.*, added_columns.data
      from base_frag
      left outer join added_columns on base_frag.id = added_columns.id)
  select * from col_adder_frag;
end;
```

You should be able to write:

```
attribute(cql:extension_fragment)
create proc adds_columns(id_ integer not null)
begin
  with
    base(*) LIKE base_frag,
    select base.*, added_columns.data
      from base
      left outer join added_columns on base.id = added_columns.id;
end;
```

Reviewed By: winniequinn

Differential Revision: D33259440

fbshipit-source-id: 823411f9cd55069f61bde8f8fafbb63b6a20163d