Generalize fetch cursor from blob to support blob expressions
Summary:
There are lots of cases where the blob is not just a local variable.  To do these cases we have to generalize the syntax, so now in addition to the simple form:

```
fetch C from blob_var;
```

There is the general form

```
fetch C from blob blob_expr;
```

The `from blob` has to be in there to avoid ambiguities, especially related to the LIKE operator on shapes.  But the simple rewrite still works so the normal cases stay the same.

Importantly, this sort of thing works:

```
fetch a_cursor from blob other_cursor.blob_field;
```

Reviewed By: winniequinn

Differential Revision: D34737276

fbshipit-source-id: 2d2efac65bb975c5ce2fb3aaabd4234a91971951