If there are any gaps in the schema versions then clear the missing facets
Summary:
In the event that a schema version "disappears" somehow we do not want any stale version CRCs in the facets table.

We can't know about forgotten recreate groups and such, they could be named anything but we can easily detect extra version CRCs and delete them.

All of this is to try to increase the chance of correctly doing some ad hoc change to the schema that isn't normally supported.  In particular, if you manually remove a bunch of stuff without leaving tombstones, and then manually delete the affected tables, you don't want vestigial CRCs in the database in case the stuff you deleted ever comes back.  So this is a little resilience, and pretty cheap.

In principle this change shouldn't be necessary but in practice people do want to make ad hoc schema changes to correct past mistakes in peculiar ways and this helps make the state a little more deterministic in the face of that sort of thing.  In short, this is to help stay sane when people are breaking the rules.

Reviewed By: winniequinn

Differential Revision: D34934766

fbshipit-source-id: f61fb86923935b064dd19f8dbbf2448fe2fd749a