@unsub/@resub get basic semantic analysis (only)
Summary:
The basic rules to unsubscribe and resubscribe are encoded here.   This is only the direct stuff that tells you if this unsub/resub even makes any sense.  The additional checks that dependent types have been properly taken care of are not yet present.

In any case, none of this *does* anything except hide the table type so far so this much is harmless.

The feature is certainly not "working" yet.

Reviewed By: winniequinn

Differential Revision: D35036293

fbshipit-source-id: e2e8af431c083bf2f48a49d6d77eb06b141f9ff1